feat(test): introduce a base class for all TransferProcessStore
