+++ core/control-plane/control-plane-core/build.gradle.kts
@@ -27,6 +27,7 @@ dependencies {
     testImplementation(testFixtures(project(":spi:common:core-spi")))
     testImplementation(testFixtures(project(":spi:control-plane:contract-spi")))
     testImplementation(testFixtures(project(":spi:control-plane:policy-spi")))
+    testImplementation(testFixtures(project(":spi:control-plane:transfer-spi")))
 }
 
 publishing {
