+++ core/control-plane/control-plane-core/src/main/java/org/eclipse/dataspaceconnector/core/controlplane/defaults/transferprocessstore/InMemoryTransferProcessStore.java
@@ -52,7 +52,7 @@ public class InMemoryTransferProcessStore implements TransferProcessStore {
 
     @Override
     @Nullable
-    public String processIdForTransferId(String id) {
+    public String processIdForDataRequestId(String id) {
         return store.findAll()
                 .filter(p -> id.equals(p.getDataRequest().getId()))
                 .findFirst()
