+++ extensions/common/iam/decentralized-identity/identity-common-test/bin/testFixtures/readme-keystore.txt
@@ -1,0 +1,1 @@
+The edc-test-keystore.jks contains a test self-signed RSA private key with the alias 'testkey'. The keystore password is test123.
