+++ extensions/control-plane/store/cosmos/transfer-process-store-cosmos/build.gradle.kts
@@ -32,6 +32,7 @@ dependencies {
     testImplementation(testFixtures(project(":extensions:common:azure:azure-test")))
     testImplementation(testFixtures(project(":extensions:common:azure:cosmos-common")))
     testImplementation("org.awaitility:awaitility:${awaitility}")
+    testImplementation(testFixtures(project(":spi:control-plane:transfer-spi")))
 }
