+++ extensions/control-plane/store/sql/transfer-process-store-sql/build.gradle.kts
@@ -36,6 +36,7 @@ dependencies {
     testImplementation("org.awaitility:awaitility:${awaitility}")
     testImplementation(testFixtures(project(":core:common:util")))
     testImplementation("org.postgresql:postgresql:${postgresVersion}")
+    testImplementation(testFixtures(project(":spi:control-plane:transfer-spi")))
 
 }
