+++ spi/control-plane/transfer-spi/build.gradle.kts
@@ -12,8 +12,14 @@
  *
  */
 
+val jupiterVersion: String by project
+val assertj: String by project
+val awaitility: String by project
+
+
 plugins {
     `java-library`
+    `java-test-fixtures`
 }
 
 
@@ -21,6 +27,12 @@ dependencies {
     api(project(":spi:common:core-spi"))
     api(project(":spi:common:policy-model"))
     api(project(":spi:common:policy-engine-spi"))
+    api(project(":core:common:policy-evaluator"))
+
+    testFixturesImplementation("org.junit.jupiter:junit-jupiter-api:${jupiterVersion}")
+    testFixturesImplementation("org.junit.jupiter:junit-jupiter-params:${jupiterVersion}")
+    testFixturesImplementation("org.assertj:assertj-core:${assertj}")
+    testFixturesImplementation("org.awaitility:awaitility:${awaitility}")
 }
 
 publishing {
